requirements.txt
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.0.4
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958435052/work
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
gitdb==4.0.9
GitPython==3.1.18
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1610099771815/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-base_1594091701667/work
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
mpmath==1.3.0
munch==3.0.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.7.0.72
pandas==1.1.5
pathtools==0.1.2
Pillow==5.4.1
pretrainedmodels==0.7.4
protobuf==3.19.6
psutil==5.9.5
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3
PyYAML==6.0
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
segmentation-models-pytorch==0.3.1
sentry-sdk==1.24.0
setproctitle==1.2.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
sympy==1.9
threadpoolctl==3.1.0
timm==0.4.12
torch==1.10.1
torchaudio==0.7.0a0+a853dff
torchvision==0.11.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094701020/work
tqdm==4.64.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
urllib3==1.26.16
wandb==0.15.3
zipp==3.6.0
